Make the app run in Release mode
I had defined GM_CHECK and GM_ASSERT as no-op in Release mode, so all the calls wrapped in them were not executed :D. In fact, all Vulkan related calls were not executed. 

It might be worth it to invest some time into proper error management and get rid of the assertions.