Fix validation errors and cleanup vertices on exit
The scissor was not set as dynamic state for the graphics pipeline and we were not freeing the device memory used by the vertex buffer (which we were also not destroying).

The validation layers don't seem to be working with the NVIDIA Quadro P5000 I have on the Shadow machine and there were no validation errors there. Need to investigate why that's the case