Introduce a Vulkan::Device struct to handle Vulkan resorces
The Renderer now is only concerned with pipelines and submitting  commands while the Vulkan::Device is responsible for handling presentation and frame resources lifetimes.

We also introduced a FrameResources struct that holds all the resources for a single frame of animation.

Another important change is that now we record commands into the command buffer every frame, and we have at most 3 frames in flight.

While refactoring I've left out the render_pass (and dependent resources) recreation on resize, but the renderer seems to be working fine (no validation errors), so I'll leave it out for now. The framebuffers are recreated for every frame