Introduce an index buffer and use the vkCmdDrawIndexed command
The vertex_buffer and index_buffer will be merged into one big buffer. The Vulkan::Device struct will allocate it and will give out offset of it to the Renderer, which will be able to copy data (vertices, indices and uniforms) into it via staging buffers.

We'll try to implement this buffer memory management as specified in this article: https://developer.nvidia.com/vulkan-memory-management. But before we do that we need to explore how Uniform Buffers work, cause they'll be part of the one big buffer.