Basic game running with WASD and keys
This commits a second graphics pipeline for rendering the player and uses push constants for sending the player matrix to the shaders.

The code to control the player and to represent the state of the game is all over the place and it's not meant to live long. We'll soon introduce a simulation struct that will carry all the simulations for the game and will produce a state that will be consumed by the renderer