Fix queues synchronization
It was sufficient to submit an Image Memory Barrier to the graphics queue to make it wait until the texture was ready to be used.

This fix got rid of the need to wait on a semaphore to be signaled, which is heavy-weight and blocks execution of any command on the graphics queue. And it was not really working :D